feat: update bell elements
