+++ packages/js/src/ui/components/index.ts
@@ -1,1 +1,1 @@
-export * from './Bell';
+export * from './BellContainer';
